+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - VIMEET-1914 - Bold on sheet titles
+### Added
+- VIMEET-1950 - In back-office, all participants are assigned to meeting if option is activated.
+- VIMEET-1931 - Unallow link if sheet has a scheduled meeting
 
 ## [2.1.0] - 2019-04-24
 ### Fixed